Improve pidTPCML.cxx
* Out of line code to reduce memory usage while compiling pidTPC tasks
* Change to use an array for the intermediate tracks representation.
  It would probably be better to read directly the values from the
  arrow tables, but that will come in a different PR.
